+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Command for publishing assets.
+- Show alert if assets are not up-to-date
 
 ## [0.23.1] - 2023-08-22